requirements.txt
argostranslate==1.9.6
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ctranslate2==4.6.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
protobuf==6.30.2
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacremoses==0.0.53
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
stanza==1.1.1
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
